+++ CHANGES.md
@@ -3,6 +3,7 @@
 * cross-compilation: use `wine` CLI when `wine64`
   is not available.
 * Add support for `ipv6only` socket option.
+* Add API to use posix sockets for connecting.
 
 0.3.1 (2024-10-29)
 =====